[Refactor] Move PCP attention operations under v1
